# apps/api_server/assets/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/products/":{"get":{"tags":["products"],"summary":"Retrieve Listing","description":"Handle GET listing request.","operationId":"retrieve_listing_api_v1_products__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"products_per_page","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Products Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WrappedProductListing"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/products/{product_id}":{"get":{"tags":["products"],"summary":"Get Product","description":"Handle GET product request.","operationId":"get_product_api_v1_products__product_id__get","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","title":"Product Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WrappedProductDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Home","description":"Handle requests for the homepage.","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Home  Get"}}}}}}},"/health":{"get":{"summary":"Health","description":"Handle health queries.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ProductDetails":{"properties":{"sku":{"type":"integer","title":"Sku"},"brand":{"type":"string","title":"Brand"},"description":{"type":"string","title":"Description"},"unit_price":{"type":"number","title":"Unit Price"},"qty_in_stock":{"type":"integer","title":"Qty In Stock"}},"type":"object","required":["sku","brand","description","unit_price","qty_in_stock"],"title":"ProductDetails","description":"Display a product's complete details."},"ProductShortInfo":{"properties":{"sku":{"type":"integer","title":"Sku"},"full_name":{"type":"string","title":"Full Name"}},"type":"object","required":["sku","full_name"],"title":"ProductShortInfo","description":"Schema to use to display a product's short info."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WrappedProductDetails":{"properties":{"data":{"$ref":"#/components/schemas/ProductDetails"}},"type":"object","required":["data"],"title":"WrappedProductDetails","description":"Schema to use to display a product's details."},"WrappedProductListing":{"properties":{"data":{"items":{"$ref":"#/components/schemas/ProductShortInfo"},"type":"array","title":"Data"},"count":{"type":"integer","title":"Count"},"previous":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Previous"},"next":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next"}},"type":"object","required":["data","count","previous","next"],"title":"WrappedProductListing","description":"Schema to use to display a product listing."}}}}